requirements.txt
# Core dependencies
asyncio-mqtt>=0.11.1
aiohttp>=3.9.1
cryptography>=41.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Database
aiosqlite>=0.19.0

# Browser automation
playwright>=1.41.0

# AI and LLM
ollama>=0.1.7
langchain>=0.1.0
langchain-community>=0.0.10

# GUI and notifications
plyer>=2.1.0
pystray>=0.19.4

# Utilities
python-dateutil>=2.8.2
pydantic>=2.5.0
click>=8.1.7
rich>=13.7.0
schedule>=1.2.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
